+++ py/desispec/io/frame.py
@@ -65,7 +65,7 @@ def write_frame(outfile, frame, header=None, fibermap=None, units=None):
 
     hdus.append( fits.ImageHDU(frame.ivar.astype('f4'), name='IVAR') )
     hdus.append( fits.CompImageHDU(frame.mask, name='MASK') )
-    hdus.append( fits.ImageHDU(frame.wave.astype('f4'), name='WAVELENGTH') )
+    hdus.append( fits.ImageHDU(frame.wave.astype('f8'), name='WAVELENGTH') )
     hdus[-1].header['BUNIT'] = 'Angstrom'
     if frame.resolution_data is not None:
         hdus.append( fits.ImageHDU(frame.resolution_data.astype('f4'), name='RESOLUTION' ) )